Initial Styling Complete
This version includes division between user rows (later this will have
to be dynamically sized to fit multiple people within a single tablet),
current month and username tag, last month tag, and working profile pic
as a button.
Next step is to properly query the database and create this from the
real dates and the real data.
Then we can think about animations that bring this a little bit more to
life-so people understand what is happening.